Revert "Hunter"
