+++ src/main/java/io/github/tkjonesy/ONNX/Yolo.java
@@ -86,8 +86,7 @@ public abstract class Yolo {
 
         // standard nms
         while (!bboxes.isEmpty()) {
-//            float[] bestBbox = bboxes.removeLast();
-            float [] bestBbox = bboxes.remove(bboxes.size() - 1);
+            float[] bestBbox = bboxes.removeLast();
             bestBboxes.add(bestBbox);
             bboxes = bboxes.stream().filter(a -> computeIOU(a, bestBbox) < nmsThreshold).collect(Collectors.toList());
         }
