+++ src/main/resources/onnxSettings.properties
@@ -3,8 +3,6 @@
 modelPath=./ai_models/yolo11m.onnx
 labelPath=./src/main/resources/coco.names
 PROCESS_EVERY_NTH_FRAME=30
-CAMERA_FRAME_RATE=30
-VIDEO_CAPTURE_DEVICE_ID=0
 
 confThreshold=0.3
 nmsThreshold=0.4
